add regex to col('..').exclude('..')
