+++ polars/polars-lazy/src/test.rs
@@ -1913,3 +1913,15 @@ fn test_agg_exprs() -> Result<()> {
     assert_eq!(Vec::from(&out), &[Some(4), Some(1)]);
     Ok(())
 }
+
+#[test]
+fn test_exclude_regex() -> Result<()> {
+    let df = fruits_cars();
+    let out = df
+        .lazy()
+        .select([col("*").exclude("^(fruits|cars)$")])
+        .collect()?;
+
+    assert_eq!(out.get_column_names(), &["A", "B"]);
+    Ok(())
+}
